All enums install

+++ lib/swig/java/Egg_Groups.java
@@ -4,27 +4,27 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Egg_Groups
 {
-    public static final int NONE = 0;
-    public static final int MONSTER = 1;
-    public static final int WATER1 = 2;
-    public static final int BUG = 3;
-    public static final int FLYING = 4;
-    public static final int GROUND = 5;
-    public static final int FAIRY = 6;
-    public static final int PLANT = 7;
-    public static final int HUMANSHAPE = 8;
-    public static final int WATER3 = 9;
-    public static final int MINERAL = 10;
-    public static final int INDETERMINATE = 11;
-    public static final int WATER2 = 12;
-    public static final int DITTO = 13;
-    public static final int DRAGON = 14;
-    public static final int NO_EGGS = 15;
+    public static final long NONE = 0;
+    public static final long MONSTER = 1;
+    public static final long WATER1 = 2;
+    public static final long BUG = 3;
+    public static final long FLYING = 4;
+    public static final long GROUND = 5;
+    public static final long FAIRY = 6;
+    public static final long PLANT = 7;
+    public static final long HUMANSHAPE = 8;
+    public static final long WATER3 = 9;
+    public static final long MINERAL = 10;
+    public static final long INDETERMINATE = 11;
+    public static final long WATER2 = 12;
+    public static final long DITTO = 13;
+    public static final long DRAGON = 14;
+    public static final long NO_EGGS = 15;
 }
