+++ lib/swig/java/Genders.java
@@ -4,14 +4,14 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Genders
 {
-    public static final int FEMALE = 0;
-    public static final int MALE = 1;
-    public static final int GENDERLESS = 2;
+    public static final long FEMALE = 0;
+    public static final long MALE = 1;
+    public static final long GENDERLESS = 2;
 }
