+++ lib/swig/java/Markings.java
@@ -4,17 +4,17 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Markings
 {
-    public static final int CIRCLE = 0;
-    public static final int TRIANGLE = 1;
-    public static final int SQUARE = 2;
-    public static final int HEART = 3;
-    public static final int STAR = 4;
-    public static final int DIAMOND = 5;
+    public static final long CIRCLE = 0;
+    public static final long TRIANGLE = 1;
+    public static final long SQUARE = 2;
+    public static final long HEART = 3;
+    public static final long STAR = 4;
+    public static final long DIAMOND = 5;
 }
