+++ lib/swig/java/Move_Classes.java
@@ -4,15 +4,15 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Move_Classes
 {
-    public static final int NONE = 0;
-    public static final int STATUS = 1;
-    public static final int PHYSICAL = 2;
-    public static final int SPECIAL = 3;
+    public static final long NONE = 0;
+    public static final long STATUS = 1;
+    public static final long PHYSICAL = 2;
+    public static final long SPECIAL = 3;
 }
