+++ lib/swig/java/Natures.java
@@ -4,37 +4,37 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Natures
 {
-    public static final int NONE = 0;
-    public static final int HARDY = 1;
-    public static final int BOLD = 2;
-    public static final int MODEST = 3;
-    public static final int CALM = 4;
-    public static final int TIMID = 5;
-    public static final int LONELY = 6;
-    public static final int DOCILE = 7;
-    public static final int MILD = 8;
-    public static final int GENTLE = 9;
-    public static final int HASTY = 10;
-    public static final int ADAMANT = 11;
-    public static final int IMPISH = 12;
-    public static final int BASHFUL = 13;
-    public static final int CAREFUL = 14;
-    public static final int RASH = 15;
-    public static final int JOLLY = 16;
-    public static final int NAUGHTY = 17;
-    public static final int LAX = 18;
-    public static final int QUIRKY = 19;
-    public static final int NAIVE = 20;
-    public static final int BRAVE = 21;
-    public static final int RELAXED = 22;
-    public static final int QUIET = 23;
-    public static final int SASSY = 24;
-    public static final int SERIOUS = 25;
+    public static final long NONE = 0;
+    public static final long HARDY = 1;
+    public static final long BOLD = 2;
+    public static final long MODEST = 3;
+    public static final long CALM = 4;
+    public static final long TIMID = 5;
+    public static final long LONELY = 6;
+    public static final long DOCILE = 7;
+    public static final long MILD = 8;
+    public static final long GENTLE = 9;
+    public static final long HASTY = 10;
+    public static final long ADAMANT = 11;
+    public static final long IMPISH = 12;
+    public static final long BASHFUL = 13;
+    public static final long CAREFUL = 14;
+    public static final long RASH = 15;
+    public static final long JOLLY = 16;
+    public static final long NAUGHTY = 17;
+    public static final long LAX = 18;
+    public static final long QUIRKY = 19;
+    public static final long NAIVE = 20;
+    public static final long BRAVE = 21;
+    public static final long RELAXED = 22;
+    public static final long QUIET = 23;
+    public static final long SASSY = 24;
+    public static final long SERIOUS = 25;
 }
