+++ lib/swig/java/Stats.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Stats
 {
-    public static final int NONE = 0;
-    public static final int HP = 1;
-    public static final int ATTACK = 2;
-    public static final int DEFENSE = 3;
-    public static final int SPECIAL_ATTACK = 4;
-    public static final int SPECIAL_DEFENSE = 5;
-    public static final int SPEED = 6;
-    public static final int ACCURACY = 7;
-    public static final int EVASION = 8;
-    public static final int SPECIAL = 9;
+    public static final long NONE = 0;
+    public static final long HP = 1;
+    public static final long ATTACK = 2;
+    public static final long DEFENSE = 3;
+    public static final long SPECIAL_ATTACK = 4;
+    public static final long SPECIAL_DEFENSE = 5;
+    public static final long SPEED = 6;
+    public static final long ACCURACY = 7;
+    public static final long EVASION = 8;
+    public static final long SPECIAL = 9;
 }
