+++ lib/swig/java/Types.java
@@ -4,32 +4,32 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Types
 {
-    public static final int NONE = 0;
-    public static final int NORMAL = 1;
-    public static final int FIGHTING = 2;
-    public static final int FLYING = 3;
-    public static final int POISON = 4;
-    public static final int GROUND = 5;
-    public static final int ROCK = 6;
-    public static final int BUG = 7;
-    public static final int GHOST = 8;
-    public static final int STEEL = 9;
-    public static final int FIRE = 10;
-    public static final int WATER = 11;
-    public static final int GRASS = 12;
-    public static final int ELECTRIC = 13;
-    public static final int PSYCHIC = 14;
-    public static final int ICE = 15;
-    public static final int DRAGON = 16;
-    public static final int DARK = 17;
-    public static final int FAIRY = 18;
-    public static final int QUESTION_MARK = 10001;
-    public static final int SHADOW = 10002;
+    public static final long NONE = 0;
+    public static final long NORMAL = 1;
+    public static final long FIGHTING = 2;
+    public static final long FLYING = 3;
+    public static final long POISON = 4;
+    public static final long GROUND = 5;
+    public static final long ROCK = 6;
+    public static final long BUG = 7;
+    public static final long GHOST = 8;
+    public static final long STEEL = 9;
+    public static final long FIRE = 10;
+    public static final long WATER = 11;
+    public static final long GRASS = 12;
+    public static final long ELECTRIC = 13;
+    public static final long PSYCHIC = 14;
+    public static final long ICE = 15;
+    public static final long DRAGON = 16;
+    public static final long DARK = 17;
+    public static final long FAIRY = 18;
+    public static final long QUESTION_MARK = 10001;
+    public static final long SHADOW = 10002;
 }
