+++ lib/swig/java/Version_Groups.java
@@ -4,27 +4,27 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Version_Groups
 {
-    public static final int NONE = 0;
-    public static final int RED_BLUE = 1;
-    public static final int YELLOW = 2;
-    public static final int GOLD_SILVER = 3;
-    public static final int CRYSTAL = 4;
-    public static final int RUBY_SAPPHIRE = 5;
-    public static final int EMERALD = 6;
-    public static final int FIRERED_LEAFGREEN = 7;
-    public static final int DIAMOND_PEARL = 8;
-    public static final int PLATINUM = 9;
-    public static final int HEARTGOLD_SOULSILVER = 10;
-    public static final int BLACK_WHITE = 11;
-    public static final int COLOSSEUM = 12;
-    public static final int XD = 13;
-    public static final int BLACK_2_WHITE_2 = 14;
-    public static final int X_Y = 15;
+    public static final long NONE = 0;
+    public static final long RED_BLUE = 1;
+    public static final long YELLOW = 2;
+    public static final long GOLD_SILVER = 3;
+    public static final long CRYSTAL = 4;
+    public static final long RUBY_SAPPHIRE = 5;
+    public static final long EMERALD = 6;
+    public static final long FIRERED_LEAFGREEN = 7;
+    public static final long DIAMOND_PEARL = 8;
+    public static final long PLATINUM = 9;
+    public static final long HEARTGOLD_SOULSILVER = 10;
+    public static final long BLACK_WHITE = 11;
+    public static final long COLOSSEUM = 12;
+    public static final long XD = 13;
+    public static final long BLACK_2_WHITE_2 = 14;
+    public static final long X_Y = 15;
 }
