+++ lib/swig/java/Versions.java
@@ -4,36 +4,36 @@
  * Distributed under the MIT License (MIT) (See accompanying file LICENSE.txt
  * or copy at http://opensource.org/licenses/MIT)
  *
- * This file was generated: 2014-07-09 20:00:52.757753
+ * This file was generated: 2014-07-09 20:44:30.941573
  */
 
 package org.nc;
 
 public class Versions
 {
-    public static final int NONE = 0;
-    public static final int RED = 1;
-    public static final int BLUE = 2;
-    public static final int YELLOW = 3;
-    public static final int GOLD = 4;
-    public static final int SILVER = 5;
-    public static final int CRYSTAL = 6;
-    public static final int RUBY = 7;
-    public static final int SAPPHIRE = 8;
-    public static final int EMERALD = 9;
-    public static final int FIRERED = 10;
-    public static final int LEAFGREEN = 11;
-    public static final int DIAMOND = 12;
-    public static final int PEARL = 13;
-    public static final int PLATINUM = 14;
-    public static final int HEARTGOLD = 15;
-    public static final int SOULSILVER = 16;
-    public static final int BLACK = 17;
-    public static final int WHITE = 18;
-    public static final int COLOSSEUM = 19;
-    public static final int XD = 20;
-    public static final int BLACK_2 = 21;
-    public static final int WHITE_2 = 22;
-    public static final int X = 23;
-    public static final int Y = 24;
+    public static final long NONE = 0;
+    public static final long RED = 1;
+    public static final long BLUE = 2;
+    public static final long YELLOW = 3;
+    public static final long GOLD = 4;
+    public static final long SILVER = 5;
+    public static final long CRYSTAL = 6;
+    public static final long RUBY = 7;
+    public static final long SAPPHIRE = 8;
+    public static final long EMERALD = 9;
+    public static final long FIRERED = 10;
+    public static final long LEAFGREEN = 11;
+    public static final long DIAMOND = 12;
+    public static final long PEARL = 13;
+    public static final long PLATINUM = 14;
+    public static final long HEARTGOLD = 15;
+    public static final long SOULSILVER = 16;
+    public static final long BLACK = 17;
+    public static final long WHITE = 18;
+    public static final long COLOSSEUM = 19;
+    public static final long XD = 20;
+    public static final long BLACK_2 = 21;
+    public static final long WHITE_2 = 22;
+    public static final long X = 23;
+    public static final long Y = 24;
 }
